User harici database tablolarinin endpoint apileri yazildi.Controller ile test edildi.

+++ Models/FavoriteRecipe.cs
@@ -9,7 +9,7 @@ public partial class FavoriteRecipe
 
     public int? UserId { get; set; }
 
-    public short? RecipeId { get; set; }
+    public int? RecipeId { get; set; }
 
     public virtual Recipe? Recipe { get; set; }
